Add functions for validating static websites, fixing imports, and managing JavaScript dependencies REFACTO
